Removing level load and save functions
It seems that my save / load system doesn't work at all when trying it in a webgl version.
I got error saying I can't access the drive, also in the LevelSelection script, I am using function (UnityEditor) that can be used only in the editor, so functions to add onClick Event on button in real time doesn't seems to work.

So I am going to remove all the features that allow to make levels and load them within the game. So I removed the script MapStartBuilder from the scene and I just copy pasted all the gameobject loaded into the scene.